Remove 'GetPythonInterpreterMemoryUsage' from the 'python' package
Since Pympler segfautl under Python3, memory usage is only available
when compile with the 'cpython' build flag (until we found a replacement
for Pympler).
