+++ cmd/agent/api/agent/python_status.go
@@ -1,0 +1,23 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2019 Datadog, Inc.
+
+// Package agent implements the api endpoints for the `/agent` prefix.
+// This group of endpoints is meant to provide high-level functionalities
+// at the agent level.
+
+// +build !cpython
+
+package agent
+
+import (
+	"net/http"
+
+	"github.com/DataDog/datadog-agent/pkg/util/log"
+)
+
+func getPythonStatus(w http.ResponseWriter, r *http.Request) {
+	log.Warnf("Error: python stats is only available when compiled with cpython\n")
+	w.Write([]byte("{}")) // return empty json
+}
