+++ pkg/collector/python/helpers.go
@@ -127,177 +127,6 @@ func getModuleName(modulePath string) string {
 	return toks[len(toks)-1]
 }
 
-//// GetPythonInterpreterMemoryUsage collects python interpreter memory usage
-//func GetPythonInterpreterMemoryUsage() ([]*PythonStats, error) {
-//	glock := newStickyLock()
-//	defer glock.unlock()
-//
-//	memStatModule := python.PyImport_ImportModule(pyMemModule)
-//	if memStatModule == nil {
-//		return nil, fmt.Errorf("Unable to import Python module: %s", pyMemModule)
-//	}
-//	defer memStatModule.DecRef()
-//
-//	statter := memStatModule.GetAttrString(pyMemSummaryFunc)
-//	if statter == nil {
-//		pyErr, err := glock.getPythonError()
-//
-//		if err != nil {
-//			return nil, fmt.Errorf("An error occurred while grabbing the python memory statter: %v", err)
-//		}
-//		return nil, errors.New(pyErr)
-//	}
-//	defer statter.DecRef()
-//
-//	args := python.PyTuple_New(0)
-//	kwargs := python.PyDict_New()
-//	defer args.DecRef()
-//	defer kwargs.DecRef()
-//
-//	stats := statter.Call(args, kwargs)
-//	if stats == nil {
-//		pyErr, err := glock.getPythonError()
-//
-//		if err != nil {
-//			return nil, fmt.Errorf("An error occurred collecting python memory stats: %v", err)
-//		}
-//		return nil, errors.New(pyErr)
-//	}
-//	defer stats.DecRef()
-//
-//	keys := python.PyDict_Keys(stats)
-//	if keys == nil {
-//		pyErr, err := glock.getPythonError()
-//
-//		if err != nil {
-//			return nil, fmt.Errorf("An error occurred collecting python memory stats: %v", err)
-//		}
-//		return nil, errors.New(pyErr)
-//	}
-//	defer keys.DecRef()
-//
-//	myPythonStats := []*PythonStats{}
-//	var entry *python.PyObject
-//	for i := 0; i < python.PyList_GET_SIZE(keys); i++ {
-//		entryName := python.PyString_AsString(python.PyList_GetItem(keys, i))
-//		entry = python.PyDict_GetItemString(stats, entryName)
-//		if entry == nil {
-//			pyErr, err := glock.getPythonError()
-//
-//			if err != nil {
-//				log.Warnf("An error occurred while iterating the memory entry : %v", err)
-//			} else {
-//				log.Warnf("%v", pyErr)
-//			}
-//
-//			continue
-//		}
-//
-//		n := python.PyDict_GetItemString(entry, "n")
-//		if n == nil {
-//			pyErr, err := glock.getPythonError()
-//
-//			if err != nil {
-//				log.Warnf("An error occurred while iterating the memory entry : %v", err)
-//			} else {
-//				log.Warnf("%v", pyErr)
-//			}
-//
-//			continue
-//		}
-//
-//		sz := python.PyDict_GetItemString(entry, "sz")
-//		if sz == nil {
-//			pyErr, err := glock.getPythonError()
-//
-//			if err != nil {
-//				log.Warnf("An error occurred while iterating the memory entry : %v", err)
-//			} else {
-//				log.Warnf("%v", pyErr)
-//			}
-//
-//			continue
-//		}
-//
-//		pyStat := &PythonStats{
-//			Type:     entryName,
-//			NObjects: python.PyInt_AsLong(n),
-//			Size:     python.PyInt_AsLong(sz),
-//			Entries:  []*PythonStatsEntry{},
-//		}
-//
-//		entries := python.PyDict_GetItemString(entry, "entries")
-//		if entries == nil {
-//			continue
-//		}
-//
-//		for i := 0; i < python.PyList_GET_SIZE(entries); i++ {
-//			ref := python.PyList_GetItem(entries, i)
-//			if ref == nil {
-//				pyErr, err := glock.getPythonError()
-//
-//				if err != nil {
-//					log.Warnf("An error occurred while iterating the entry details: %v", err)
-//				} else {
-//					log.Warnf("%v", pyErr)
-//				}
-//
-//				continue
-//			}
-//
-//			obj := python.PyList_GetItem(ref, 0)
-//			if obj == nil {
-//				pyErr, err := glock.getPythonError()
-//
-//				if err != nil {
-//					log.Warnf("An error occurred while iterating the entry details : %v", err)
-//				} else {
-//					log.Warnf("%v", pyErr)
-//				}
-//
-//				continue
-//			}
-//
-//			nEntry := python.PyList_GetItem(ref, 1)
-//			if nEntry == nil {
-//				pyErr, err := glock.getPythonError()
-//
-//				if err != nil {
-//					log.Warnf("An error occurred while iterating the entry details : %v", err)
-//				} else {
-//					log.Warnf("%v", pyErr)
-//				}
-//
-//				continue
-//			}
-//
-//			szEntry := python.PyList_GetItem(ref, 2)
-//			if szEntry == nil {
-//				pyErr, err := glock.getPythonError()
-//
-//				if err != nil {
-//					log.Warnf("An error occurred while iterating the entry details : %v", err)
-//				} else {
-//					log.Warnf("%v", pyErr)
-//				}
-//
-//				continue
-//			}
-//
-//			pyEntry := &PythonStatsEntry{
-//				Reference: python.PyString_AsString(obj),
-//				NObjects:  python.PyInt_AsLong(nEntry),
-//				Size:      python.PyInt_AsLong(szEntry),
-//			}
-//			pyStat.Entries = append(pyStat.Entries, pyEntry)
-//		}
-//
-//		myPythonStats = append(myPythonStats, pyStat)
-//	}
-//
-//	return myPythonStats, nil
-//}
-
 // GetPythonIntegrationList collects python datadog installed integrations list
 func GetPythonIntegrationList() ([]string, error) {
 	glock := newStickyLock()
